Added the Analog Clock for Windows as an example.
I have found a relatively simple way to modify `analogClock.aec` so that it can run on 32-bit Windows, and saved it as `analogClockForWindows.aec`. The `analogClockForWindows.aec`, unfortunately, can only be assembled using the version of GNU Assembler that comes with MinGW-W64, and not using the one that comes with TDM-GCC or the one that comes with CygWin.